refactor:修改GameSrv绿毒点数最达65535点以及清理不必要的转换

+++ src/GameSrv/Actor/BaseObject.Base.cs
@@ -348,7 +348,7 @@ namespace GameSrv.Actor
                     {
                         ((AnimalObject)this).MeatQuality -= 1000;
                     }
-                    DamageHealth((ushort)(GreenPoisoningPoint + 1));
+                    DamageHealth(GreenPoisoningPoint + 1);
                     HealthTick = 0;
                     SpellTick = 0;
                     HealthSpellChanged();
